+++ README.md
@@ -27,7 +27,7 @@ tracking_tracksycle/
 │   │   ├── __init__.py
 │   │   ├── track.py
 │   │   ├── motion_model.py
-│   │   └── Track_Length.py
+│   │   └── track_length.py
 │   ├── playback/
 │   │   ├── __init__.py
 │   │   └── set_playhead.py
